removed redundnat code, moved models to model folder for complaints
