--- backend/src/types/complaints/allegation_search_options.ts
@@ -1,10 +1,0 @@
-export interface AllegationSearchOptions {
-    community: string;
-    zone: string;
-    region: string;
-    officerAssigned: string;
-    violationCode: string;
-    incidentReportedStart: string;
-    incidentReportedEnd: string;
-    status: string;
-  }
