--- backend/src/types/complaints/hwcr_search_options.ts
@@ -1,11 +1,0 @@
-export interface HWCRSearchOptions {
-    community: string;
-    zone: string;
-    region: string;
-    officerAssigned: string;
-    natureOfComplaint: string;
-    speciesCode: string;
-    incidentReportedStart: Date;
-    incidentReportedEnd: Date;
-    status: string;
-  }
