--- backend/src/types/complaints/map-search-results.ts
@@ -1,9 +1,0 @@
-import { AllegationComplaint } from "../../v1/allegation_complaint/entities/allegation_complaint.entity";
-import { HwcrComplaint } from "../../v1/hwcr_complaint/entities/hwcr_complaint.entity";
-import { WildlifeComplaintDto } from "../models/complaints/wildlife-complaint";
-import { AllegationComplaintDto } from "../models/complaints/allegation-complaint";
-
-export interface MapSearchResults {
-    complaints: HwcrComplaint[] | AllegationComplaint[] | WildlifeComplaintDto[] | AllegationComplaintDto[];
-    unmappedComplaints: number;
-  }
