+++ backend/src/types/models/complaints/map-search-results.ts
@@ -1,0 +1,7 @@
+import { WildlifeComplaintDto } from "./wildlife-complaint";
+import { AllegationComplaintDto } from "./allegation-complaint";
+
+export interface MapSearchResults {
+  complaints: WildlifeComplaintDto[] | AllegationComplaintDto[];
+  unmappedComplaints: number;
+}
