+++ backend/src/types/models/complaints/update-complaint.dto.ts
@@ -1,13 +1,13 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { AgencyCode } from '../../agency_code/entities/agency_code.entity';
-import { ComplaintStatusCode } from '../../complaint_status_code/entities/complaint_status_code.entity';
-import { GeoOrganizationUnitCode } from '../../geo_organization_unit_code/entities/geo_organization_unit_code.entity';
+import { AgencyCode } from '../../../v1/agency_code/entities/agency_code.entity';
+import { ComplaintStatusCode } from '../../../v1/complaint_status_code/entities/complaint_status_code.entity';
+import { GeoOrganizationUnitCode } from '../../../v1/geo_organization_unit_code/entities/geo_organization_unit_code.entity';
 import { Point } from 'geojson';
-import { CosGeoOrgUnit } from '../../cos_geo_org_unit/entities/cos_geo_org_unit.entity';
-import { PersonComplaintXref } from '../../person_complaint_xref/entities/person_complaint_xref.entity';
+import { CosGeoOrgUnit } from '../../../v1/cos_geo_org_unit/entities/cos_geo_org_unit.entity';
+import { PersonComplaintXref } from '../../../v1/person_complaint_xref/entities/person_complaint_xref.entity';
 import { ReportedByCode } from 'src/v1/reported_by_code/entities/reported_by_code.entity';
 
-export class ComplaintDto {
+export class UpdateComplaintDto {
   @ApiProperty({
     example: "COS-3425",
     description: "The identifier of the complaint",
