--- backend/src/v1/complaint/dto/update-complaint.dto.ts
@@ -1,3 +1,0 @@
-import { CreateComplaintDto } from './create-complaint.dto';
-
-export class UpdateComplaintDto extends CreateComplaintDto {}
